requirements.txt
streamlit
pandas
requests
snowflake-connector-python